auth

+++ app/Http/Controllers/PostController.php
@@ -4,8 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -14,8 +13,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        // $posts = DB::table("posts")->select("id", "title", "content", "created_at")->where("active", true)->get();
-        // $posts = Post::where("active", true)->get();
+        if (!Auth::check()) {
+            return redirect("login")->with("error_message", "Silakan login dahulu");
+        }
 
         // menampilkan data yang sudah soft deletes
         // $posts = Post::active()->withTrashed()->get();
@@ -23,6 +23,7 @@ class PostController extends Controller
         $view_data = [
             "posts" => $posts
         ];
+        // return response($view_data);
         return view('posts.index', $view_data);
     }
 
@@ -31,6 +32,10 @@ class PostController extends Controller
      */
     public function create()
     {
+        if (!Auth::check()) {
+            return redirect("login")->with("error_message", "Silakan login dahulu");
+        }
+
         return view("posts.create");
     }
 
@@ -39,21 +44,13 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect("login")->with("error_message", "Silakan login dahulu");
+        }
+
         $title = $request->input("title");
         $content = $request->input("content");
 
-        // DB::table("posts")->insert([
-        //     "title" => $title,
-        //     "content" => $content,
-        //     "created_at" => date("Y-m-d H:i:s"),
-        //     "updated_at" => date("Y-m-d H:i:s")
-        // ]);
-        // Post::insert([
-        //     "title" => $title,
-        //     "content" => $content,
-        //     "created_at" => date("Y-m-d H:i:s"),
-        //     "updated_at" => date("Y-m-d H:i:s")
-        // ]);
         Post::create([
             "title" => $title,
             "content" => $content
@@ -67,6 +64,10 @@ class PostController extends Controller
      */
     public function show(string $id)
     {
+        if (!Auth::check()) {
+            return redirect("login")->with("error_message", "Silakan login dahulu");
+        }
+
         // $post = DB::table("posts")->select("id", "title", "content", "created_at")->where("id", "=", $id)->first();
         $post = Post::where("id", $id)->first();
         $comments = $post->comments()->get();
@@ -85,6 +86,10 @@ class PostController extends Controller
      */
     public function edit(string $id)
     {
+        if (!Auth::check()) {
+            return redirect("login")->with("error_message", "Silakan login dahulu");
+        }
+
         // $post = DB::table("posts")->select("id", "title", "content", "created_at")->where("id", "=", $id)->first();
         $post = Post::where("id", $id)->first();
 
@@ -100,6 +105,10 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (!Auth::check()) {
+            return redirect("login")->with("error_message", "Silakan login dahulu");
+        }
+
         $title = $request->input("title");
         $content = $request->input("content");
 
@@ -122,6 +131,10 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        if (!Auth::check()) {
+            return redirect("login")->with("error_message", "Silakan login dahulu");
+        }
+
         // DB::table("posts")->where("id", "=", $id)->delete();
         Post::where("id", $id)->delete();
